chore: Update user model to include psychologist and patients fields
This commit updates the user model in the backend to include the psychologist and patients fields. The psychologist field is of type mongoose.Schema.Types.ObjectId and is used to reference the User model, allowing users to be associated with a psychologist. The patients field is an array of mongoose.Schema.Types.ObjectId and is used to reference multiple User models, representing the patients associated with a psychologist. This change enhances the functionality of the application by providing a way to track the relationship between users and psychologists.