chore: Remove unnecessary code and update user role handling
This commit removes unnecessary code in the backend Seed.js file that was previously used to clear existing data. Additionally, it updates the role handling in the backend and frontend codebase. The superadminController has been renamed to adminController, and the middleware functions have been updated accordingly. This change improves code clarity and maintainability.