refactor: Remove unused import in BlogCard.js and update image alt text in BlogPreviewModal.js
This commit refactors the code in BlogCard.js to remove the unused import of FaEdit from react-icons/fa. It also updates the image alt text in BlogPreviewModal.js to use a more generic alt text for the blog images. These changes improve the code cleanliness and accessibility of the application.